Don't dictate how trial results are accumulated in monte carlo sims
Problem
-------
When we run monte carlo simulations, we need to accumulate the results.

Imagine having 5,000,000 results -- you don't want that in memory!

But, somtimes you *do* want the trial results, if they are smaller
and you have a use for them.

Previously, we solved this by using Kixi's histogram function.
What that really does, under the hood, is to use Tad Dunning's
T-Digest data structure.

Essentially, this "compresses" the data so you can still get accurate
quantiles without storing everything in memory.

However -- sometimes we do and sometimes we don't need the trial
results. And, Kixi's histogram function is Java-only, whereas most
of this library is cljc.

Solution
--------
Introduce a protocol, TrialResultsAccumulator. Let the caller decide
how to compress the results of the trials.

This makes testing easier too!

Then, where we currently use the monte carlo runner, set the kixi
histogram as a default, but allow the caller to pass in their own
TrialResultsAccumulator.

In other words: be open for extension, closed for modification